Refactor TemplatesDirectory component structure and imports
