+++ src/app.postcss
@@ -9,7 +9,3 @@
 .textarea {
     @apply border-neutral-300 focus-within:border-neutral-300 focus:border-neutral-500;
 }
-
-input[type="file"] {
-    @apply border-none file:variant-soft-primary hover:file:variant-filled-primary !important;
-}
